Making a few changes relevant to me.
I was looking for a solution to this. I'm the original author of the script this was based on. I'm pleased to see someone went an refined the code. I'm honestly a newbie at bash scripting. more so when i cobbled that together. And about 50% of that was code i found while on Google. You happened to have solved an issue i was having with this.